before deployment test

+++ credit/models.py
@@ -5,7 +5,28 @@ from accounts.models import CustomUser
 from django.core.validators import MinValueValidator, MaxValueValidator
 from django.core.exceptions import ValidationError
 from datetime import datetime
+from dateutil.relativedelta import relativedelta # type: ignore
+import datetime
+from django.core.validators import MinValueValidator
+from django.core.files.storage import default_storage
+from django.core.exceptions import ObjectDoesNotExist
+from django.db.models import Sum
+from django.core.exceptions import ValidationError
 
+from notification.services import NotificationService
+from notification.models import Notification
+
+def notify_applicant(credit, heading, body, link):
+    """
+    Notify Credit Committee members about a new credit request that passed guarantor approval.
+    """
+    NotificationService.send_notification(
+        credit.applicant,
+        heading,
+        body,
+        link,
+        notification_type=Notification.NotificationType.IN_APP
+    )
 
 def validate_positive(value):
     if value < Decimal(0):
@@ -112,7 +133,6 @@ class CreditSettings(models.Model):
         verbose_name = "Credit Setting"
         verbose_name_plural = "Credit Settings"
 
-import datetime
 
 class Credit(models.Model):
     CREDIT_TYPE_CHOICES = [
@@ -157,6 +177,39 @@ class Credit(models.Model):
         """A credit can be canceled only if it is still pending."""
         return self.status == 'Pending'
     
+
+    @property
+    def get_credit_type_model(self):
+        """
+        Returns the specific credit type model instance associated with this credit.
+        """
+        if self.credit_type == 'Qard Hasan':
+            return None
+        elif self.credit_type == 'Murabaha':
+            return self.murabaha # type: ignore
+        elif self.credit_type == 'Musharaka':
+            return self.musharaka # type: ignore
+        elif self.credit_type == 'Ijarah':
+            return self.ijarah # type: ignore
+        else:
+            raise ValueError(f"Unknown credit type: {self.credit_type}")
+    
+    @property
+    def get_remaining_months(self):
+        """Calculate months remaining for repayment."""
+        repayment_start_date = self.repayment_start_month_as_date
+        if repayment_start_date:
+            months_elapsed = (datetime.datetime.now().year - repayment_start_date.year) * 12 + datetime.datetime.now().month - repayment_start_date.month
+            if months_elapsed < 0:
+                months_elapsed = 0
+            return self.repayment_period - months_elapsed
+        return self.repayment_period
+    
+    @property
+    def total_repaid(self):
+        """Calculate the total amount repaid for this credit."""
+        return self.repayment_set.aggregate(total=Sum('amount'))['total'] or Decimal(0) # type: ignore
+    
     @property
     def repayment_start_month_as_date(self):
         """Return the repayment start month as a date object (year, month)."""
@@ -167,6 +220,9 @@ class Credit(models.Model):
 
     def save(self, *args, **kwargs):
         """Ensure credit meets policy settings before saving."""
+        
+        if self.pk and self.credit_type == 'Murabaha':
+            self.amount_requested = self.murabaha.selling_price # type: ignore
 
         # Fetch active credit settings
         settings = CreditSettings.objects.first()
@@ -206,35 +262,54 @@ class Guarantor(models.Model):
     status = models.CharField(max_length=10, choices=status_choices, default='Pending')
     action_date = models.DateTimeField(null=True, blank=True)
 
-from django.core.files.storage import default_storage
-import os
+
+def validate_file_extension(value):
+    valid_extensions = ['.pdf']
+    if not any(value.name.lower().endswith(ext) for ext in valid_extensions):
+        raise ValidationError(f"Unsupported file extension. Allowed extension is: {', '.join(valid_extensions)}")
 
 class Murabaha(models.Model):
-    credit = models.OneToOneField(Credit, on_delete=models.CASCADE, related_name='murabaha')
+    credit = models.OneToOneField("Credit", on_delete=models.CASCADE, related_name="murabaha")
     asset_name = models.CharField(max_length=255)
-    asset_value = models.DecimalField(max_digits=12, decimal_places=2, validators=[validate_positive])
+    asset_price = models.DecimalField(max_digits=12, decimal_places=2, validators=[validate_positive])
+    profit_margin_percentage = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
+    profit_margin_fixed = models.DecimalField(max_digits=12, decimal_places=2, null=True, blank=True)  
     profit_margin = models.DecimalField(max_digits=12, decimal_places=2, validators=[validate_positive], blank=True, null=True)
-    vendor_invoice = models.ImageField(upload_to='murabaha/vendor_invoice/', max_length=100, blank=True, null=True)
+    selling_price = models.DecimalField(max_digits=12, decimal_places=2, blank=True, null=True)
+    vendor_invoice = models.FileField(upload_to="murabaha/vendor_invoice/", max_length=100, blank=True, null=True, validators=[validate_file_extension])
     created_at = models.DateField(auto_now=True)
 
+    def calculate_selling_price(self):
+        """
+        Calculate and return the selling price.
+        """
+        if self.profit_margin_fixed is not None:
+            return Decimal(self.asset_price) + Decimal(self.profit_margin_fixed)
+        elif self.profit_margin_percentage is not None:
+            return Decimal(self.asset_price) * (1 + Decimal(self.profit_margin_percentage) / 100)
+        return None
+
     def total_credit_amount(self):
-        """Calculate total cost including profit margin"""
-        if self.profit_margin is None:
-            return self.asset_value  # No profit margin
-        return self.asset_value + self.profit_margin
+        """
+        Calculate the total amount including profit margin.
+        """
+        return self.selling_price or self.asset_price  # Use selling price if already set
 
     def save(self, *args, **kwargs):
-        # If the vendor_invoice is being updated (not null) and a new file is uploaded
-        if self.pk:  # Check if the instance already exists (i.e., it's being updated)
-            existing = Murabaha.objects.get(pk=self.pk)  # Fetch existing instance
-            if self.vendor_invoice != existing.vendor_invoice:  # Check if the invoice has changed
-                # Delete the old file from the storage
-                if existing.vendor_invoice:
-                    old_invoice_path = existing.vendor_invoice.path
-                    if default_storage.exists(old_invoice_path):
-                        default_storage.delete(old_invoice_path)
+        # Auto-calculate selling price before saving
+        self.selling_price = self.calculate_selling_price()
+        
+        if self.selling_price:
+            self.profit_margin = Decimal(self.selling_price) - Decimal(self.asset_price)
 
-        super(Murabaha, self).save(*args, **kwargs)  # Call the parent class save method
+        if self.pk:
+            existing = self.__class__.objects.filter(pk=self.pk).first()
+            if existing and existing.vendor_invoice and self.vendor_invoice != existing.vendor_invoice:
+                old_invoice_path = existing.vendor_invoice.path
+                if default_storage.exists(old_invoice_path):
+                    default_storage.delete(old_invoice_path)
+
+        super().save(*args, **kwargs)
 
 class Musharaka(models.Model):
     credit = models.OneToOneField(Credit, on_delete=models.CASCADE, related_name='musharaka')
@@ -251,11 +326,36 @@ class Ijarah(models.Model):
 
 class Repayment(models.Model):
     credit = models.ForeignKey(Credit, on_delete=models.CASCADE)
-    repayment_date = models.DateTimeField()
+    repayment_date = models.DateTimeField(auto_now_add=True)
     amount = models.DecimalField(max_digits=12, decimal_places=2)
-    remaining_balance = models.DecimalField(max_digits=12, decimal_places=2)
     repayment_method = models.CharField(max_length=1, choices=[('S', 'Salary Deduction'), ('D', 'Direct Deposit')], default='S')
     status = models.CharField(max_length=20, choices=[('Repaid', 'Paid'), ('Pending', 'Pending'), ('Overdue', 'Overdue')])
+
+    
+    def save(self, *args, **kwargs):
+        # Get all existing repayments for this credit
+        # Aggregate the amount of existing repayments using SQL
+        total_repaid = self.credit.total_repaid
+        
+        
+        # Add the current amount to the total repaid
+        total_repaid += self.amount
+        
+        
+        # Update status based on remaining balance
+        if self.credit.amount_requested is not None and total_repaid >= (self.credit.amount_requested - Decimal('1.00')):
+            self.status = 'Repaid'
+            self.credit.status = 'Repaid'
+            self.credit.save()
+            heading = "Credit Repayment Completion Notification"
+            body = f"Dear {self.credit.applicant.get_full_name()}, congratulations! You have completed the repayment of your credit."
+            # link = reverse('credit:transactions')
+            link=''
+            notify_applicant(self.credit, heading, body, link)
+        else:
+            self.status = 'Pending'
+        
+        super().save(*args, **kwargs)
 
 class TransactionLog(models.Model):
     TRANSACTION_TYPE_CHOICES = [
@@ -268,10 +368,37 @@ class TransactionLog(models.Model):
     transaction_receipt = models.FileField(upload_to='transaction_receipts/', max_length=255, blank=True, null=True)
     credit = models.ForeignKey(Credit, on_delete=models.CASCADE)
 
+    def save(self, *args, **kwargs):
+        if self.transaction_type == 'R':
+            credit = self.credit
+            if credit.repayment_start_month and credit.repayment_period:
+                year, month = map(int, credit.repayment_start_month.split('-'))
+                repayment_end_date = datetime.date(year, month, 1) + relativedelta(months=credit.repayment_period)
+                if credit.credit_type == 'Murabaha':
+                    total_requested = credit.murabaha.selling_price # type: ignore
+                else:
+                    total_requested = credit.amount_requested or Decimal(0)
+                total_paid = credit.total_repaid
+                total_paid += self.amount
+                remaining_balance = total_requested - total_paid
+                remaining_months = (repayment_end_date.year - datetime.date.today().year) * 12 + (repayment_end_date.month - datetime.date.today().month)
+                if remaining_months > 0:
+                    credit.monthly_deduction = remaining_balance / remaining_months
+                else:
+                    credit.monthly_deduction = remaining_balance
+                credit.save()
+        elif self.transaction_type == 'R-S':
+            self.transaction_type = 'R'
 
+        if self.pk:
+            existing = self.__class__.objects.filter(pk=self.pk).first()
+            if existing and existing.transaction_receipt and self.transaction_receipt != existing.transaction_receipt:
+                old_receipt_path = existing.transaction_receipt.path
+                if default_storage.exists(old_receipt_path):
+                    default_storage.delete(old_receipt_path)
 
+        super().save(*args, **kwargs)
 
-from django.core.exceptions import ObjectDoesNotExist
 
 class CreditCommittee(models.Model):
     member = models.OneToOneField(CustomUser, on_delete=models.CASCADE, related_name='credit_committee_member')
